systemd: fail if "env" file is missing
The env file is essential for GNOME integration.

Before, the "-" before the path would cause sway to start but emit to
error, although something was badly wrong.

If for some reason you want a different behavior, you can use standard
system methods to override this configuration detail in the file or
the whole file.

Fixes #5