feat: Add gnome-keyring user service
The gnome-keyring user service initializes the gnome-keyring-daemon,
which is a daemon for ssh,gpg,etc.

This service is not enabled by default so `systemctl --user enable
gnome-keyring.service` must be run in order to start the keyring when
sway is started.

The dependency order of sway-session-pre was also changed beacause it
allowed wayland-session-pre.target to be started after sway.service.